+++ cached_path/schemes/scheme_client.py
@@ -8,7 +8,11 @@ class SchemeClient:
     A client used for caching remote resources corresponding to URLs with a particular scheme.
 
     Subclasses must define the :attr:`scheme` class variable and implement
-    :meth:`get_etag()` and :meth:`get_resource`.
+    :meth:`get_etag()` and :meth:`get_resource()`.
+
+    .. important::
+        Take care when implementing subclasses to raise the right error types
+        from :meth:`get_etag()` and :meth:`get_resource()`.
     """
 
     connection_error_types: ClassVar[Tuple[BaseException, ...]] = (
@@ -30,11 +34,43 @@ class SchemeClient:
     def get_etag(self) -> Optional[str]:
         """
         Get the Etag or an equivalent version identifier associated with the resource.
+
+        Returns
+        -------
+        ``Optional[str]``
+            The ETag as a ``str`` or ``None`` if there is no ETag associated with
+            the resource.
+
+        Raises
+        ------
+        ``FileNotFoundError``
+            If the resource doesn't exist.
+
+        ``Connection error``
+            Any error type defined in ``SchemeClient.connection_error_types`` will
+            be treated as a retriable connection error.
+
+        ``Other errors``
+            Any other error type can be raised, in which case ``cached_path()`` will
+            log the error and move on to try to fetch the resource without the ETag.
         """
         raise NotImplementedError
 
     def get_resource(self, temp_file: IO) -> None:
         """
         Download the resource to the given temporary file.
+
+        Raises
+        ------
+        ``FileNotFoundError``
+            If the resource doesn't exist.
+
+        ``Connection error``
+            Any error type defined in ``SchemeClient.connection_error_types`` will
+            be treated as a retriable connection error.
+
+        ``Other errors``
+            Any other error type can be raised, in which case ``cached_path()`` will
+            fail and propogate the error.
         """
         raise NotImplementedError